Criando testes específicos para Indexer
IN-2554
