+++ tests/test_log_indexer.py
@@ -171,41 +171,3 @@ class LogIndexerTest(asynctest.TestCase):
             prepared_document = self.indexer._prepare_document(self.logmessage_parse_error)
             self.assertEqual(64, prepared_document['asgard_index_delay'])
 
-    async def test_generated_right_action_data_for_bulk_insert(self):
-        """
-        Antes de passar o(s) documento(s) para o elasticsearch, devemos
-        chamar `self._prepare_document(document)`.
-        E para gerar o nome do índice onde esse documento será indexado, chamamos
-        `self._index_name(document)`
-        """
-        class MyIndexer(Indexer):
-
-            def _prepare_document(self, document):
-                del document['key']
-                document['new-key'] = 42
-                return document
-
-            def _index_name(self, document):
-                return "my-important-index"
-
-        self.elasticsearch_mock.bulk.return_value = 42
-        indexer = MyIndexer(self.elasticsearch_mock)
-        returned_by_elasticsearch = await indexer.bulk([{"key": "errors.asgard.app.sieve.captura.kirby.powerup", "some-value": 10}])
-        self.assertEqual([mock.call([
-            { "index" : { "_index" : "my-important-index", "_type" : "logs"}},
-            {"some-value": 10, "new-key": 42},
-        ])], self.elasticsearch_mock.bulk.await_args_list)
-        self.assertEqual(42, returned_by_elasticsearch)
-
-    async def test_confirms_default_prepare_document_implementation(self):
-        class MyOtherIndexer(Indexer):
-            def _index_name(self, document):
-                return "my-important-index"
-
-        indexer = MyOtherIndexer(self.elasticsearch_mock)
-        await indexer.bulk([{"key": "errors.asgard.app.sieve.captura.kirby.powerup", "some-value": 10}])
-        self.assertEqual([mock.call([
-            { "index" : { "_index" : "my-important-index", "_type" : "logs"}},
-            {"key": "errors.asgard.app.sieve.captura.kirby.powerup", "some-value": 10},
-        ])], self.elasticsearch_mock.bulk.await_args_list)
-
